backend/requirements.txt
fastapi==0.116.1
uvicorn==0.33.0
sqlalchemy==2.0.43
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
openai==1.58.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
